requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1.2
django-json-response==1.1.5
entrypoints==0.4
executing==1.1.1
ez-setup==0.9
gunicorn==20.1.0
idna==3.4
ipykernel==6.16.1
ipython==8.5.0
ipywidgets==8.0.2
jedi==0.18.1
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-widgets==3.0.3
matplotlib-inline==0.1.6
mecab-python3==1.0.5
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.3
psycopg2==2.9.4
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
tornado==6.2
traitlets==5.5.0
tzdata==2022.5
unidic-lite==1.0.8
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==6.2.0
widgetsnbextension==4.0.3